requirements.txt
streamlit==1.13.0
tensorflow==2.17.0
matplotlib==3.9.2
pillow==10.4.0
numpy==1.26.4
altair==4.2.0
pandas==2.2.2
mne==1.4.0
